Convert structure to a Blender Extension
- Add manifest
- Removed bl_info from __init__.py
- Separated JSBSim parser class into its own file
